created a crude delete page in case we need to remove content. Right now it selects the id and deletes from there and should not run if there is no id. no links to this page, remove must be entered in manually

+++ routes/apiRoutes.js
@@ -59,14 +59,23 @@ module.exports = app => {
     });
 
     app.delete("/api/:id", (req, res) => {
-        db.Image.destroy({
-            where: {
-                id: req.params.id
+        let id = req.params.id;
+        db.Image.findAndCountAll({}).then(response => {
+            if (id < response.count) {
+                db.Image.destroy({
+                    where: {
+                        id: id
+                    }
+                }).then(response => {
+                    console.log(response);
+                    res.json(response);
+                });
+            } else {
+                throw "error";
             }
-        }).then(response => {
-            console.log(response);
-            res.json(success);
         });
+
+
     });
